+++ chaosgcp/lb/__init__.py
@@ -1,12 +1,18 @@
 import logging
 import re
+from typing import List, Tuple
+from urllib.parse import urlparse
 
 from chaoslib.exceptions import ActivityFailed
 from google.cloud.compute_v1.types import compute
 
 logger = logging.getLogger("chaostoolkit")
 
-__all__ = ["get_fault_injection_policy", "remove_fault_injection_policy"]
+__all__ = [
+    "get_fault_injection_policy",
+    "remove_fault_injection_policy",
+    "get_route_action_from_url",
+]
 
 
 def get_fault_injection_policy(
@@ -21,6 +27,13 @@ def remove_fault_injection_policy(
 ) -> None:
     route_action = get_route_action(urlmap, target_name, target_path)
     route_action.fault_injection_policy = None
+
+
+def get_fault_injection_policy_from_url(
+    urlmaps: List[compute.UrlMap],
+    url: str,
+) -> Tuple[compute.UrlMap, compute.HttpRouteAction]:
+    return get_route_action_from_url(urlmaps, url)
 
 
 ###############################################################################
@@ -88,3 +101,39 @@ def get_route_action(
     logger.debug(f"Found path '{target_path}' in '{target_name}'")
 
     return route_action
+
+
+def get_route_action_from_url(
+    urlmaps: List[compute.UrlMap], url: str
+) -> Tuple[compute.UrlMap, compute.HttpRouteAction]:
+    p = urlparse(url)
+    domain = p.netloc
+
+    found_urlmap = None
+    for urlmap in urlmaps:
+        if found_urlmap:
+            break
+
+        for host_rule in urlmap.host_rules:
+            if domain in host_rule.hosts:
+                found_urlmap = urlmap
+                break
+
+    for pm in found_urlmap.path_matchers:
+        for rr in pm.route_rules:
+            for mr in rr.match_rules:
+                if mr.prefix_match is not None:
+                    pattern = re.compile(f"^{mr.prefix_match}")
+                    print(pattern, p.path, pattern.match(p.path))
+                    if pattern.match(p.path):
+                        return (found_urlmap, rr.route_action)
+                elif mr.full_path_match is not None:
+                    pattern = re.compile(f"^{mr.full_path_match}$")
+                    if pattern.match(p.path):
+                        return (found_urlmap, rr.route_action)
+                elif mr.regex_match is not None:
+                    pattern = re.compile(f"^{mr.regex_match}$")
+                    if pattern.match(p.path):
+                        return (found_urlmap, rr.route_action)
+
+    raise ActivityFailed("failed to find a suitable route")